Split classes- babel config- created bundle
