+++ 02-OOP-inheritance/index.html
@@ -8,6 +8,6 @@
 </head>
 <body>
   
-  <script src="main.js"></script>
+  <script  src="scripts/script-compiled.js" ></script>
 </body>
 </html>
