+++ 02-OOP-inheritance/scripts/lib/models/actor.js
@@ -1,0 +1,17 @@
+'use strict';
+
+Object.defineProperty(exports, "__esModule", {
+  value: true
+});
+exports.default = void 0;
+
+class Actor {
+  constructor(name, age) {
+    this.name = name;
+    this.age = age;
+  }
+
+}
+
+var _default = Actor;
+exports.default = _default;
