+++ 02-OOP-inheritance/scripts/lib/models/logger.js
@@ -1,0 +1,18 @@
+'use strict';
+
+Object.defineProperty(exports, "__esModule", {
+  value: true
+});
+exports.default = void 0;
+
+class Logger {
+  constructor() {}
+
+  log(info) {
+    console.log(info);
+  }
+
+}
+
+var _default = Logger;
+exports.default = _default;
