+++ 02-OOP-inheritance/scripts/src/models/actor.js
@@ -1,0 +1,9 @@
+'use strict';
+
+class Actor {
+    constructor(name, age) {
+        this.name = name;
+        this.age = age;
+    }
+}
+export default Actor;
