+++ 02-OOP-inheritance/scripts/src/models/logger.js
@@ -1,0 +1,10 @@
+'use strict';
+
+class Logger {
+    constructor(){}
+    
+    log(info) {
+        console.log(info);
+    }
+}
+export default Logger;
