+++ 02-OOP-inheritance/scripts/src/models/social.js
@@ -1,0 +1,14 @@
+'use strict';
+
+class Social {
+    initializer(){}
+
+    share(friendName){
+        console.log(friendName + " shares " + this.title);
+    }
+
+    like(friendName){
+        console.log(friendName + " likes " + this.title);
+    }
+}
+export default Social;
